Refactor vanilla anticheat handling + cleanup

+++ src/main/java/ac/grim/grimac/predictionengine/movementtick/MovementTickerStrider.java
@@ -25,7 +25,7 @@ public class MovementTickerStrider extends MovementTickerRideable {
     }
 
     public static boolean isAbove(GrimPlayer player) {
-        return player.y > Math.floor(player.y) + 0.5 - (double) 1.0E-5F;
+        return player.y > Math.floor(player.y) + 0.5 - 1.0E-5F;
     }
 
     @Override
